Crosswalk data for Jung et al. (2020) (level 2 classification)

The International Union for Conservation of Nature (IUCN) Red List of Threatened Species provides a habitat classification scheme for different habitat types. To map the spatial location of each of these habitat classes, Jung et al. (2020a) developed a global raster dataset by combining a range of different land use and land cover data (Jung et al. 2020b). They created two habitat classification schemes – termed level 1 and level 2 classification schemes – to describe coarse-scale and fine-scale differences in habitat classes. This dataset provides a crosswalk table to associate the values in the level 2 raster dataset with IUCN habitat classes.

Usage

data(crosswalk_jung_lvl2_data)

Format

A data frame (tibble::tibble()) object with 126 rows and 2 columns. Each row corresponds to a different IUCN habitat class. It has the following columns:

code

The character code for a given IUCN habitat class.

value

The numeric value assigned to grid cells in the raster data that contain the IUCN habitat class (see get_jung_lvl2_habitat_data()).

Source

The data were obtained from Jung et al. (2020b).

References

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V, Rondinini C, and Visconti P (2020a) A global map of terrestrial habitat types. Scientific Data, 7, 1–8. doi:10.1038/s41597-020-00599-8

Jung M, Dahal PR, Butchart SHM, Donald PF, De Lamo X, Lesiv M, Kapos V, Rondinini C, and Visconti P (2020b) A global map of terrestrial habitat types (insert version) [Data set]. Zenodo. doi:10.5281/zenodo.4058819

See also

Examples

# load data
data(crosswalk_jung_lvl2_data)

# print data
print(crosswalk_jung_lvl2_data)
#> # A tibble: 126 × 2
#>    code  value
#>    <chr> <int>
#>  1 1       100
#>  2 1.1     101
#>  3 1.2     102
#>  4 1.3     103
#>  5 1.4     104
#>  6 1.5     105
#>  7 1.6     106
#>  8 1.7     107
#>  9 1.8     108
#> 10 1.9     109
#> # ℹ 116 more rows