Default IUCN Red List Habitat Classification Scheme omission codes

The International Union for Conservation of Nature (IUCN) Red List of Threatened Species provides a habitat classification scheme for different habitat types worldwide. Under this scheme, each habitat type is assigned to a code. This function returns the codes for habitat types that correspond to artificial, unknown, miscellaneous, or invasive vegetation habitat types that should be omitted from Area of Habitat calculations.

default_omit_iucn_habitat_codes()

Value

A character vector of habitat codes.

Examples

# print omission codes
print(default_omit_iucn_habitat_codes())
#>  [1] "14"    "14.1"  "14.2"  "14.3"  "14.4"  "14.5"  "14.6"  "15"    "15.1" 
#> [10] "15.2"  "15.3"  "15.4"  "15.5"  "15.6"  "15.7"  "15.8"  "15.9"  "15.10"
#> [19] "15.11" "15.12" "15.13" "16"    "17"    "18"