Get species habitat data

Import species habitat preference data obtained from the International Union for Conservation of Nature (IUCN) Red List of Threatened Species (https://www.iucnredlist.org/).

get_spp_habitat_data(
  x,
  dir = tempdir(),
  version = "latest",
  key = NULL,
  delay = 2,
  force = FALSE,
  verbose = TRUE
)

Arguments

x
integer Taxon identifier for the species on the International Union for Conservation of Nature (IUCN) Red List of Threatened Species. For example, the taxon identifier for the species Loxodonta africana is 181008073.

dir
character folder path where data should be downloaded. By default, data are downloaded to a temporary directory (i.e. tempdir()). To avoid downloading the same data multiple times, it is strongly recommended to specify a persistent storage location (see Examples below). If data are already available in folder for the specified version, then the data are imported and not re-downloaded from the IUCN Red List.

version
character indicating version of the IUCN Red List for obtaining data (e.g. "2021-2"). Although it is not possible to query past versions of the IUCN Red List, this functionality is useful for accessing data previously downloaded from the IUCN Red List. Defaults to "latest" such that data are downloaded from the most recent version of the IUCN Red List.

key
character Token for querying the IUCN Red List API. Please note that a token is required to download data from the IUCN Red List. To obtain a token, please see rredlist::rl_use_iucn(). Defaults to NULL such that the token accessed from the "IUCN_REDLIST_KEY" environmental variable (which can be specified in the .Renviron file).

delay
integer number of seconds to wait between subsequent calls to the IUCN Red List API. Defaults to 2 seconds (as recommended by the rredlist package; https://docs.ropensci.org/rredlist/articles/rredlist.html).

force
logical should the data be downloaded even if the the data are already available? Defaults to FALSE.

verbose
logical should progress be displayed while downloading data? Defaults to TRUE.

Value

A table (tibble::tibble()) object.

Details

Data are downloaded from the IUCN Red List using the rredlist::rl_habitats(). This function is essentially a wrapper designed to help download data for multiple species and provide caching for previously downloaded data.

References

Please use the following citation for data obtained from the IUCN Red List:

IUCN (insert year). IUCN Red List of Threatened Species. Version

To obtain the version number of the latest version, use rredlist::rl_version().

Examples

# \dontrun{
# define species to download data for based on taxon identifiers
spp_ids <- c(18, 22694927)

# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# download and import habitat preference data
spp_habitat_data <- get_spp_habitat_data(spp_ids, download_dir)
# }

if (FALSE) { # interactive()
# \dontrun{
# preview data
print(result)
# }
}