Intersecting extent

Generate a terra::ext() object containing the spatial extent of one terra::rast() raster object inside another terra::rast() object.

intersecting_ext(x, y, buffer = 0)

Arguments

x
terra::rast() Raster object. This object specifies the terra::crs() coordinate reference system used for the output object.

y
terra::rast() Raster object. This object specifies the spatial extent for the output object.

buffer
numeric buffer applied to the spatial extent of y before reprojecting to the coordinate reference system of x. Defaults to 0.

Value

A terra::ext() object.

See also

Examples

# define raster object
x <- rast(
  ncols = 40, nrows = 40,
  xmin = -110, xmax = -90, ymin = 40, ymax = 60,
  crs = "+proj=longlat +datum=WGS84"
)
print(x)
#> class       : SpatRaster 
#> dimensions  : 40, 40, 1  (nrow, ncol, nlyr)
#> resolution  : 0.5, 0.5  (x, y)
#> extent      : -110, -90, 40, 60  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +datum=WGS84 +no_defs 

# define raster object with different coordinate reference system
y <- rast(
  ncols = 94, nrows = 124,
  xmin = -944881, xmax = 935118, ymin = 4664377, ymax = 7144377,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)
print(y)
#> class       : SpatRaster 
#> dimensions  : 124, 94, 1  (nrow, ncol, nlyr)
#> resolution  : 19999.99, 20000  (x, y)
#> extent      : -944881, 935118, 4664377, 7144377  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=lcc +lat_0=0 +lon_0=-100 +lat_1=48 +lat_2=33 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 

# find spatial extent of x that intersects with the spatial extent of y
print(intersecting_ext(x, y))
#> SpatExtent : -110, -90, 40, 60 (xmin, xmax, ymin, ymax)