Parallel project

Project a terra::rast() object to a conform to the spatial properties of another terra::rast() object. Additionally, parallel processing can be used to reduce run time.

parallel_project(
  x,
  y,
  method = "bilinear",
  buffer = 0,
  parallel_n_threads = 1,
  parallel_cluster = NULL,
  temp_dir = tempdir(),
  verbose = TRUE,
  ...
)

Arguments

x
terra::rast() Raster object with data needs to be projected.

y
terra::rast() Raster object with spatial properties that are used to define the output (i.e. spatial extent, resolution, and coordinate reference system).

method
character Name for projection method. See terra::project() for available options. Defaults to "bilinear".

buffer
numeric Buffer for clipping data. Defaults to 0.

parallel_n_threads
integer Number of computational threads to use for data processing. To reduce run time, it is strongly recommended to set this parameter based on the number of available threads (see Examples below). Defaults to 1.

parallel_cluster
character Name of strategy for processing data in parallel. Available options are "FORK" and "PSOCK". Defaults to NULL such that "PSOCK" is used on Microsoft Windows operating systems, and "FORK" otherwise. Defaults to 1.

temp_dir
character Directory for saving temporary files. Defaults to base::tempdir().

verbose
logical Should progress be displayed while processing data? Defaults to TRUE.

...
arguments passed to terra::crop() and terra::project().

Value

A terra::ext() object.

Details

This function speeds up data projection by manually excluding data from x that do not overlap with the spatial extent of y (accounting for a pre-specified buffer).

Examples

# define raster object with data for projection
x <- rast(
  ncols = 40, nrows = 40,
  xmin = -110, xmax = -90, ymin = 40, ymax = 60,
  crs = "+proj=longlat +datum=WGS84"
)
x <- rast(lapply(seq_len(10), function(i) {
 init(x, runif)
}))

# preview object
print(x)
#> class       : SpatRaster 
#> dimensions  : 40, 40, 10  (nrow, ncol, nlyr)
#> resolution  : 0.5, 0.5  (x, y)
#> extent      : -110, -90, 40, 60  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +datum=WGS84 +no_defs 
#> sources     : memory  
#>               memory  
#>               memory  
#>               ... and 7 more source(s)
#> names       :        lyr.1,        lyr.1,        lyr.1,        lyr.1,        lyr.1,        lyr.1, ... 
#> min values  : 1.036914e-04, 2.173744e-04, 5.742875e-04, 1.741570e-03, 4.762653e-04, 1.154277e-04, ... 
#> max values  :    0.9996525,    0.9990620,    0.9987542,    0.9996459,    0.9998727,    0.9997498, ... 

# define raster object with spatial properties used to perform
# the projection processing
y <- rast(
  ncols = 94, nrows = 124,
  xmin = -944881, xmax = 935118, ymin = 4664377, ymax = 7144377,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)
print(y)
#> class       : SpatRaster 
#> dimensions  : 124, 94, 1  (nrow, ncol, nlyr)
#> resolution  : 19999.99, 20000  (x, y)
#> extent      : -944881, 935118, 4664377, 7144377  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=lcc +lat_0=0 +lon_0=-100 +lat_1=48 +lat_2=33 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 

# project data using a single thread
z <- parallel_project(x, y, parallel_n_threads = 1)
print(z)
#> class       : SpatRaster 
#> dimensions  : 124, 94, 10  (nrow, ncol, nlyr)
#> resolution  : 19999.99, 20000  (x, y)
#> extent      : -944881, 935118, 4664377, 7144377  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=lcc +lat_0=0 +lon_0=-100 +lat_1=48 +lat_2=33 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> sources     : memory  
#>               memory  
#>               memory  
#>               ... and 7 more source(s)
#> names       :       lyr.1,       lyr.1,       lyr.1,       lyr.1,       lyr.1,       lyr.1, ... 
#> min values  : 0.020958189, 0.022002839, 0.005753596, 0.019276554, 0.012947499, 0.013497977, ... 
#> max values  :   0.9811843,   0.9891549,   0.9740787,   0.9755028,   0.9976288,   0.9882557, ... 

# project data using multiple threads
# \dontrun{
z2 <- parallel_project(x, y, parallel_n_threads = 1)
print(z2)
#> class       : SpatRaster 
#> dimensions  : 124, 94, 10  (nrow, ncol, nlyr)
#> resolution  : 19999.99, 20000  (x, y)
#> extent      : -944881, 935118, 4664377, 7144377  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=lcc +lat_0=0 +lon_0=-100 +lat_1=48 +lat_2=33 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> sources     : memory  
#>               memory  
#>               memory  
#>               ... and 7 more source(s)
#> names       :       lyr.1,       lyr.1,       lyr.1,       lyr.1,       lyr.1,       lyr.1, ... 
#> min values  : 0.020958189, 0.022002839, 0.005753596, 0.019276554, 0.012947499, 0.013497977, ... 
#> max values  :   0.9811843,   0.9891549,   0.9740787,   0.9755028,   0.9976288,   0.9882557, ... 
# }