Disaggregate a raster using GDAL

This function is a wrapper for gdalUtils::gdal_translate().

terra_gdal_disaggregate(
  x,
  fact,
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  datatype = "FLT4S",
  tiled = FALSE,
  verbose = TRUE,
  ...
)

Arguments

x
terra::rast() Raster object with source data.

fact
integer factor to disaggregate data.

n_threads
integer Number of computational threads to use for data processing. To reduce run time, it is strongly recommended to set this parameter based on available resources (see Examples section below). Defaults to 1.

filename
character Filename for output raster. Defaults to tempfile(fileext = ".tif").

datatype
character Value indicating the data type for saving data. Defaults to "FLT4S".

tiled
logical Value indicating if GeoTIFF files should be tiled. Defaults to FALSE.

verbose
logical should information be displayed during processing? Defaults to TRUE.

...
Arguments passed to terra::writeRaster().

Value

A terra::rast() raster object.

Examples

# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

# create raster with data
x <- rast(
  ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
  crs = "+proj=longlat +datum=WGS84"
)
values(x) <- seq_len(ncell(x))

# disaggregate data
z <- terra_gdal_disaggregate(x, fact = 10)
#> Checking gdal_installation...
#> GDAL version 3.2.2
#> GDAL command being used: "/usr/bin/gdal_translate" -tr 0.05 0.05 -ot "Float32" -of "GTiff" -co "COMPRESS=LZW" -co "NUM_THREADS=1" -oo "NUM_THREADS=1" "/tmp/RtmpVpJt9L/filea341b6cebddb5.tif" "/tmp/RtmpVpJt9L/filea341b2e51f39d.tif"
#> Input file size is 40, 400...10...20...30...40...50...60...70...80...90...100 - done.

# preview result
print(z)
#> class       : SpatRaster 
#> dimensions  : 400, 400, 1  (nrow, ncol, nlyr)
#> resolution  : 0.05, 0.05  (x, y)
#> extent      : -110, -90, 40, 60  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source      : filea341b2e51f39d.tif 
#> name        : lyr.1