Convert vector data into a raster using GDAL

This function is a wrapper for gdalUtils::gdal_rasterize().

terra_gdal_rasterize(
  x,
  sf,
  burn = 1,
  invert = FALSE,
  update = FALSE,
  n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  sf_filename = tempfile(fileext = ".gpkg"),
  datatype = "FLT4S",
  tiled = FALSE,
  bigtiff = FALSE,
  compress = "LZW",
  verbose = TRUE,
  output_raster = TRUE
)

Arguments

x
terra::rast() Raster object with source data.

sf
sf::st_sf() Spatial object to rasterize.

burn
numeric Value for encoding the vector data. Defaults to 1.

invert
logical Should the burn process be inverted? Defaults to FALSE.

update
logical Should the result by producing by updating the argument to x? If FALSE then the argument to x is only used to specify the spatial properties of the resulting raster (i.e. values have on the result), Defaults to FALSE.

n_threads
integer Number of computational threads to use for data processing. To reduce run time, it is strongly recommended to set this parameter based on available resources (see Examples section below). Defaults to 1.

filename
character Filename for output raster. Defaults to tempfile(fileext = ".tif").

sf_filename
character File name to temporarily save argument to sf. Defaults to a temporary (geopackage) file.

datatype
character Value indicating the data type for saving data. Defaults to "FLT4S".

tiled
logical Value indicating if GeoTIFF files should be tiled. Defaults to FALSE.

bigtiff
logical Value indicating the data should be stored in BIGTIFF format. Defaults to FALSE.

compress
character Value indicating compression format. Defaults to "LZW".

verbose
logical Should information be displayed during processing? Defaults to TRUE.

output_raster
logical Should a raster (terra::rast()) object be returned? If FALSE then the file path for the resulting file is returned. Defaults to TRUE.

Value

A terra::rast() raster object.

Examples

# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

# TODO