IUCN habitat classification codes

The International Union for Conservation of Nature (IUCN) Red List of Threatened Species provides a habitat classification scheme for different habitat types worldwide. This dataset provides details on the different IUCN habitat classes.

data(iucn_habitat_data)

Format

A data frame with 126 rows and 2 columns. Each row corresponds to a different habitat class, and each column contains information about a given habitat class. It contains columns with the following values for each habitat class.

code

A character value indicating the code of a class.

name

A character value indicating the name of a class.

is_terrestrial

A logical value indicating if the class occurs within the terrestrial environmental.

is_marine

A logical value indicating if the class occurs within the marine environmental.

is_artificial

A logical value indicating if the class is artificial (e.g., anthropogenically human modified).

is_misc

A logical value indicating if the class does not correspond to a specific land cover.

is_introduced

A logical value indicating if the class corresponds to introduced vegetation.

Source

The data were obtained from https://www.iucnredlist.org/resources/habitat-classification-scheme.

Examples

# load data
data(iucn_habitat_data)

# print data
print(iucn_habitat_data)
#> # A tibble: 126 × 7
#>    code  name                            is_te…¹ is_ma…² is_ar…³ is_misc is_in…⁴
#>    <chr> <chr>                           <lgl>   <lgl>   <lgl>   <lgl>   <lgl>  
#>  1 1     Forest                          TRUE    FALSE   FALSE   FALSE   FALSE  
#>  2 1.1   Forest - Boreal                 TRUE    FALSE   FALSE   FALSE   FALSE  
#>  3 1.2   Forest - Subarctic              TRUE    FALSE   FALSE   FALSE   FALSE  
#>  4 1.3   Forest - Subantarctic           TRUE    FALSE   FALSE   FALSE   FALSE  
#>  5 1.4   Forest - Temperate              TRUE    FALSE   FALSE   FALSE   FALSE  
#>  6 1.5   Forest - Subtropical/tropical … TRUE    FALSE   FALSE   FALSE   FALSE  
#>  7 1.6   Forest - Subtropical/tropical … TRUE    FALSE   FALSE   FALSE   FALSE  
#>  8 1.7   Forest - Subtropical/tropical … TRUE    FALSE   FALSE   FALSE   FALSE  
#>  9 1.8   Forest - Subtropical/tropical … TRUE    FALSE   FALSE   FALSE   FALSE  
#> 10 1.9   Forest - Subtropical/tropical … TRUE    FALSE   FALSE   FALSE   FALSE  
#> # … with 116 more rows, and abbreviated variable names ¹​is_terrestrial,
#> #   ²​is_marine, ³​is_artificial, ⁴​is_introduced