Simulate species data

Simulate species data for creating Area of Habitat data (Brooks et al. 2019). Specifically, data are simulated to define species geographic ranges, summary information, and habitat preferences.

simulate_spp_data(
  n,
  boundary_data,
  habitat_data = NULL,
  elevation_data = NULL,
  rf_model_scale_min = 1e+05,
  rf_model_scale_max = 2e+05,
  cache_dir = tempdir(),
  habitat_version = "latest",
  force = FALSE,
  omit_habitat_codes = default_omit_iucn_habitat_codes(),
  verbose = TRUE
)

Arguments

n
integer Number of species to simulate.

boundary_data
sf::st_sf() Spatial object delineating the spatial extent and boundary for simulating species ranges.

habitat_data
terra::rast() Multi-layer raster data delineating the coverage of different habitat classes across world (e.g. Jung et al. 2020a). Each layer should correspond to a different habitat class, and the name of each layer should contain a unique code used to identify which places contain suitable habitat for a given species (per the "code" column in the the argument to spp_habitat_data). For each layer, pixel values should indicate the percentage of available an available habitat class. Critically, a value of 1000 should indicate 100%, and a value of 0 should indicate 0%. Defaults to NULL such that data are automatically obtained from Jung et al. 2020b (using get_global_habitat_data()).

elevation_data
terra::rast() Raster data delineating the worldwide elevation data (e.g. Jung et al. 2020a). Defaults to NULL such that data are automatically obtained from Jung et al. 2020b (using get_global_elevation_data()).

rf_model_scale_min
numeric Minimum scaling parameter used to control the smallest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 1e+5.

rf_model_scale_max
numeric Maximum scaling parameter used to control the largest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 2e+5.

cache_dir
character folder path for downloading and caching data. By default, a temporary directory is used (i.e. tempdir()). To avoid downloading the same data multiple times, it is strongly recommended to specify a persistent storage location (see Examples below).

habitat_version
character Version of the habitat dataset that should be used. See documentation for the the version parameter in the get_global_habitat_data() function for further details. Defaults to "latest" such that the most recent version of the dataset is used.

force
logical should the data be downloaded even if the the data are already available? Defaults to FALSE.

omit_habitat_codes
character Habitat classification codes to omit from resulting Area of Habitat data. If the aim is to identify places that contain natural, suitable habitats, then processing should exclude (i) anthropogenically modified habitat classifications and (ii) unknown habitat classification types. Defaults to all artificial, introduced vegetation, and unknown habitat classifications on the IUCN Red List Habitat Classification Scheme (see default_omit_iucn_habitat_codes(),

verbose
logical Should progress be displayed while processing data? Defaults to TRUE.

Value

A list object containing simulated data that are formatted following conventions used by the International Union for Conservation of Nature (IUCN) Red List of Threatened Species. It contains the following elements:

spp_range_data

A sf::st_sf() object containing the species' geographic range data.

spp_summary_data

A tibble::tibble() object containing summary information about the species (including altitudinal limit information.

spp_habitat_data

A tibble::tibble() object containing habitat preferences for the species.

References

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W, Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V, Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat (AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution. 34:977--986. doi:10.1016/j.tree.2019.06.009

See also

Examples

# please ensure that the RandomFields and smoothr packages are installed
# to run these examples

# \dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# specify file path for boundary data
boundary_path <- system.file("shape/nc.shp", package = "sf")

# import boundary data to simulate species data
boundary_data <- sf::st_union(sf::read_sf(boundary_path))

# specify
# simulate data for 5 species
x <- simulate_spp_data(
  n = 5, boundary_data = boundary_data, cache_dir = download_dir
)
#> ℹ importing global elevation data
#> ✔ importing global elevation data ... done
#> 
#> ℹ importing global habitat data
#> ✔ importing global habitat data ... done
#> 
#> Simple feature collection with 4 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -7387088 ymin: 4154562 xmax: -7287088 ymax: 4361562
#> Projected CRS: World_Behrmann
#>     id_no                       geometry
#> 1     698 POLYGON ((-7350088 4361562,...
#> 1.1   698 POLYGON ((-7296088 4226562,...
#> 1.2   698 POLYGON ((-7311088 4219562,...
#> 1.3   698 POLYGON ((-7361088 4186562,...
#> Simple feature collection with 1 feature and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8136088 ymin: 4080562 xmax: -7477088 ymax: 4358562
#> Projected CRS: World_Behrmann
#>   id_no                       geometry
#> 1  2511 POLYGON ((-7674088 4358562,...
#> Simple feature collection with 5 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8136088 ymin: 4086562 xmax: -7283088 ymax: 4362562
#> Projected CRS: World_Behrmann
#>     id_no                       geometry
#> 1    4320 POLYGON ((-7285088 4245562,...
#> 1.1  4320 POLYGON ((-7311088 4219562,...
#> 1.2  4320 POLYGON ((-8136088 4199562,...
#> 1.4  4320 POLYGON ((-7361088 4186562,...
#> 1.5  4320 POLYGON ((-7839088 4362562,...
#> Simple feature collection with 5 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8136088 ymin: 4088562 xmax: -7280088 ymax: 4343562
#> Projected CRS: World_Behrmann
#>     id_no                       geometry
#> 1    4359 POLYGON ((-7288088 4279562,...
#> 1.1  4359 POLYGON ((-7311088 4219562,...
#> 1.2  4359 POLYGON ((-7887088 4343562,...
#> 1.4  4359 POLYGON ((-7361088 4186562,...
#> 1.5  4359 POLYGON ((-7546088 4139562,...
#> Simple feature collection with 5 features and 1 field
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8123088 ymin: 4154562 xmax: -7286088 ymax: 4364562
#> Projected CRS: World_Behrmann
#>     id_no                       geometry
#> 1.1  6710 POLYGON ((-7293088 4227562,...
#> 1.2  6710 POLYGON ((-7311088 4219562,...
#> 1.3  6710 POLYGON ((-7881088 4364562,...
#> 1.4  6710 POLYGON ((-7401088 4201562,...
#> 1.5  6710 POLYGON ((-7361088 4186562,...

# preview species range data
print(x$spp_range_data)
#> Simple feature collection with 18 features and 26 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: -84.3235 ymin: 33.88305 xmax: -75.45658 ymax: 36.58977
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    id_no seasonal presence origin                       geometry
#> 1    698        1        1      1 MULTIPOLYGON (((-76.54392 3...
#> 2    698        1        4      3 MULTIPOLYGON (((-76.01663 3...
#> 3    698        1        1      2 MULTIPOLYGON (((-75.98029 3...
#> 4    698        1        5      5 MULTIPOLYGON (((-75.74029 3...
#> 5   2511        1        1      1 POLYGON ((-84.32331 34.9897...
#> 6   4320        1        1      1 POLYGON ((-81.24197 36.5686...
#> 7   4320        1        6      4 POLYGON ((-76.54392 34.5879...
#> 8   4320        1        5      5 POLYGON ((-76.01663 35.0696...
#> 9   4320        1        3      6 POLYGON ((-75.73697 35.1928...
#> 10  4359        2        1      1 POLYGON ((-84.32295 34.9899...
#>             binomial   compiler yrcompiled citation subspecies subpop source
#> 1   Simulus spp. 698 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 2   Simulus spp. 698 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 3   Simulus spp. 698 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 4   Simulus spp. 698 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 5  Simulus spp. 2511 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 6  Simulus spp. 4320 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 7  Simulus spp. 4320 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 8  Simulus spp. 4320 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 9  Simulus spp. 4320 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 10 Simulus spp. 4359 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#>    island tax_comm dist_comm generalisd legend kingdom phylum class order_
#> 1    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 2    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 3    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 4    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 5    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 6    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 7    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 8    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 9    <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 10   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#>    family   genus category marine terrestial freshwater
#> 1    <NA> Simulus       LC  false       true      false
#> 2    <NA> Simulus       LC  false       true      false
#> 3    <NA> Simulus       LC  false       true      false
#> 4    <NA> Simulus       LC  false       true      false
#> 5    <NA> Simulus       LC  false       true      false
#> 6    <NA> Simulus       LC  false       true      false
#> 7    <NA> Simulus       LC  false       true      false
#> 8    <NA> Simulus       LC  false       true      false
#> 9    <NA> Simulus       LC  false       true      false
#> 10   <NA> Simulus       LC  false       true      false

# preview species habitat affiliation data
print(x$spp_habitat_data)
#> # A tibble: 21 × 6
#>    id_no code  habitat                      suitability season   majorimportance
#>    <int> <chr> <chr>                        <chr>       <chr>    <chr>          
#>  1   698 3     Shrubland                    Suitable    Resident NA             
#>  2   698 4     Grassland                    Suitable    Resident NA             
#>  3   698 3.4   Shrubland - Temperate        Suitable    Resident NA             
#>  4  2511 6     Rocky Areas (e.g., inland c… Suitable    Resident NA             
#>  5  2511 4.4   Grassland - Temperate        Suitable    Resident NA             
#>  6  2511 3     Shrubland                    Suitable    Resident NA             
#>  7  4320 1.4   Forest - Temperate           Suitable    Resident NA             
#>  8  4320 5.5   Wetlands (inland) - Permane… Suitable    Resident NA             
#>  9  4320 5.6   Wetlands (inland) - Seasona… Suitable    Resident NA             
#> 10  4359 1.4   Forest - Temperate           Suitable    Breedin… NA             
#> # … with 11 more rows

# preview species summary data
print(x$spp_summary_data)
#> # A tibble: 5 × 31
#>   id_no taxonid scientific_name  kingdom phylum class order family genus  
#>   <int>   <int> <chr>            <chr>   <chr>  <chr> <chr> <chr>  <chr>  
#> 1   698     698 Simulus spp. 698 NA      NA     NA    NA    NA     Simulus
#> 2  2511    2511 Simulus spp. 698 NA      NA     NA    NA    NA     Simulus
#> 3  4320    4320 Simulus spp. 698 NA      NA     NA    NA    NA     Simulus
#> 4  4359    4359 Simulus spp. 698 NA      NA     NA    NA    NA     Simulus
#> 5  6710    6710 Simulus spp. 698 NA      NA     NA    NA    NA     Simulus
#> # … with 22 more variables: main_common_name <chr>, authority <chr>,
#> #   published_year <chr>, assessment_date <chr>, category <chr>,
#> #   criteria <chr>, population_trend <chr>, marine_system <chr>,
#> #   freshwater_system <chr>, terrestrial_system <chr>, assessor <chr>,
#> #   reviewer <chr>, aoo_km2 <chr>, eoo_km2 <chr>, elevation_upper <dbl>,
#> #   elevation_lower <dbl>, depth_upper <dbl>, depth_lower <dbl>,
#> #   errata_flag <chr>, errata_reason <chr>, amended_flag <chr>, …
# }