Simulate species data

Simulate species data for creating Area of Habitat data (Brooks et al. 2019). Specifically, data are simulated to define species geographic ranges, summary information, and habitat preferences.

Usage,

simulate_spp_data(
  n,
  boundary_data,
  habitat_data = NULL,
  elevation_data = NULL,
  crosswalk_data = NULL,
  rf_model_scale_min = 1e+05,
  rf_model_scale_max = 2e+05,
  cache_dir = tempdir(),
  habitat_version = "latest",
  force = FALSE,
  omit_habitat_codes = iucn_habitat_codes_marine(),
  verbose = TRUE
)

Arguments

n

integer Number of species to simulate.

boundary_data

sf::st_sf() Spatial object delineating the spatial extent and boundary for simulating species ranges.

habitat_data

terra::rast() Raster data indicating the presence of different habitat classes across world (e.g. Jung et al. 2020a,b). Each pixel should contain an integer value that specifies which habitat class is present within the pixel (based on the argument to crosswalk_data). Defaults to NULL such that data are automatically obtained (using get_global_habitat_data()).

elevation_data

terra::rast() Raster data delineating the worldwide elevation data (e.g. Robinson et al. 2014). Defaults to NULL such that data are automatically obtained (using get_global_elevation_data()). If the data are obtained automatically, then a preprocessed version of the habitat data will be used to reduce processing time.

crosswalk_data

data.frame() Table containing data that indicate which pixel values in the argument to habitat_data correspond to which IUCN habitat classification codes. The argument should contain a code column that specifies a set of IUCN habitat classification codes (see iucn_habitat_data(), and a value column that specifies different values in the argument to habitat_data. Defaults to NULL such that the crosswalk for the default habitat data are used (i.e. crosswalk_jung_data()).

rf_model_scale_min

numeric Minimum scaling parameter used to control the smallest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 1e+5.

rf_model_scale_max

numeric Maximum scaling parameter used to control the largest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 2e+5.

cache_dir

character Folder path for downloading and caching data. By default, a temporary directory is used (i.e. tempdir()). To avoid downloading the same data multiple times, it is strongly recommended to specify a persistent storage location (see Examples below).

habitat_version

character Version of the habitat dataset that should be used. See documentation for the the version parameter in the get_global_habitat_data() function for further details. This parameter is only used if habitat data are obtained automatically (i.e. the argument to habitat_data is NULL). Defaults to "latest" such that the most recent version of the dataset is used if data need to be obtained.

force

logical Should the data be downloaded even if the the data are already available? Defaults to FALSE.

omit_habitat_codes

character Habitat classification codes to omit from resulting Area of Habitat data. Please see the IUCN Red List Habitat Classification Scheme for the full range of habitat classification codes. For example, if the aim is to identify natural places that contain suitable conditions, then areas classified as anthropogenically modified (iucn_habitat_codes_artificial()), introduced vegetation (iucn_habitat_codes_introduced(), or unknown habitat (iucn_habitat_codes_misc()) should be excluded. Defaults to iucn_habitat_codes_marine(), such that marine habitats are excluded.

verbose

logical Should progress be displayed while processing data? Defaults to TRUE.

Value

A list object containing simulated data that are formatted following conventions used by the International Union for Conservation of Nature (IUCN) Red List of Threatened Species. It contains the following elements:

spp_range_data

A sf::st_sf() object containing the species' geographic range data.

spp_summary_data

A tibble::tibble() object containing summary information about the species (including altitudinal limit information.

spp_habitat_data

A tibble::tibble() object containing habitat preferences for the species.

References

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W, Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V, Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat (AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution, 34:977--986. Available at https://doi.org/10.1016/j.tree.2019.06.009.

See also

Examples

# please ensure that the RandomFields and smoothr packages are installed
# to run these examples

# \dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# specify file path for boundary data
boundary_path <- system.file("shape/nc.shp", package = "sf")

# import boundary data to simulate species data
boundary_data <- sf::st_union(sf::read_sf(boundary_path))

# set random number generator seeds for consistency
set.seed(500)
RandomFields::RFoptions(seed = 500)

# simulate data for 5 species
x <- simulate_spp_data(
  n = 5, boundary_data = boundary_data, cache_dir = download_dir
)
#> ℹ importing global elevation data
#> ✔ importing global elevation data [11.2s]
#> 
#> ℹ importing global habitat data
#> NOTE: simulation is performed with fixed random seed 500.
#> Set 'RFoptions(seed=NA)' to make the seed arbitrary.
#> NOTE: simulation is performed with fixed random seed 500.
#> NOTE: simulation is performed with fixed random seed 500.
#> NOTE: simulation is performed with fixed random seed 500.
#> NOTE: simulation is performed with fixed random seed 500.
#> ✔ importing global habitat data [40.4s]
#> 

# preview species range data
print(x$spp_range_data)
#> Simple feature collection with 8 features and 26 fields
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: -84.32377 ymin: 33.88216 xmax: -75.45658 ymax: 36.58977
#> Geodetic CRS:  WGS 84
#>   id_no seasonal presence origin                       geometry
#> 1   799        1        1      1 MULTIPOLYGON (((-82.57872 3...
#> 2  2102        1        1      1 MULTIPOLYGON (((-82.19164 3...
#> 3  2102        1        5      5 MULTIPOLYGON (((-84.31003 3...
#> 4  4082        1        1      1 MULTIPOLYGON (((-81.9608 35...
#> 5  4082        1        3      4 MULTIPOLYGON (((-76.29354 3...
#> 6  4082        1        3      2 MULTIPOLYGON (((-77.88534 3...
#> 7  5167        1        1      1 POLYGON ((-81.97127 35.1884...
#> 8  5479        1        1      1 MULTIPOLYGON (((-79.07426 3...
#>            binomial   compiler yrcompiled citation subspecies subpop source
#> 1  Simulus spp. 799 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 2 Simulus spp. 2102 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 3 Simulus spp. 2102 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 4 Simulus spp. 4082 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 5 Simulus spp. 4082 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 6 Simulus spp. 4082 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 7 Simulus spp. 5167 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#> 8 Simulus spp. 5479 Simulation         NA     <NA>       <NA>   <NA>   <NA>
#>   island tax_comm dist_comm generalisd legend kingdom phylum class order_
#> 1   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 2   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 3   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 4   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 5   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 6   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 7   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#> 8   <NA>     <NA>      <NA>         NA   <NA>    <NA>   <NA>  <NA>   <NA>
#>   family   genus category marine terrestial freshwater
#> 1   <NA> Simulus       NT  false       true      false
#> 2   <NA> Simulus       NT  false       true      false
#> 3   <NA> Simulus       NT  false       true      false
#> 4   <NA> Simulus       NT  false       true      false
#> 5   <NA> Simulus       NT  false       true      false
#> 6   <NA> Simulus       NT  false       true      false
#> 7   <NA> Simulus       NT  false       true      false
#> 8   <NA> Simulus       NT  false       true      false

# preview species habitat affiliation data
print(x$spp_habitat_data)
#> # A tibble: 13 × 6
#>    id_no code  habitat                        suitability season majorimportance
#>    <int> <chr> <chr>                          <chr>       <chr>  <chr>          
#>  1   799 3.4   Shrubland - Temperate          Suitable    Resid… NA             
#>  2   799 5.15  Wetlands (inland) - Seasonal/… Suitable    Resid… NA             
#>  3  2102 5     Wetlands (inland)              Suitable    Resid… NA             
#>  4  2102 14.2  Pastureland                    Suitable    Resid… NA             
#>  5  2102 5.5   Wetlands (inland) - Permanent… Suitable    Resid… NA             
#>  6  4082 14.2  Pastureland                    Suitable    Resid… NA             
#>  7  4082 14.3  Plantations                    Suitable    Resid… NA             
#>  8  4082 14.5  Urban Areas                    Suitable    Resid… NA             
#>  9  5167 3.4   Shrubland - Temperate          Suitable    Resid… NA             
#> 10  5167 14.1  Arable Land                    Suitable    Resid… NA             
#> 11  5167 1.8   Forest - Subtropical/tropical… Suitable    Resid… NA             
#> 12  5479 3.4   Shrubland - Temperate          Suitable    Resid… NA             
#> 13  5479 5.13  Wetlands (inland) - Permanent… Suitable    Resid… NA             

# preview species summary data
print(x$spp_summary_data)
#> # A tibble: 5 × 31
#>   id_no taxonid scientific_name  kingdom phylum class order family genus  
#>   <int>   <int> <chr>            <chr>   <chr>  <chr> <chr> <chr>  <chr>  
#> 1   799     799 Simulus spp. 799 NA      NA     NA    NA    NA     Simulus
#> 2  2102    2102 Simulus spp. 799 NA      NA     NA    NA    NA     Simulus
#> 3  4082    4082 Simulus spp. 799 NA      NA     NA    NA    NA     Simulus
#> 4  5167    5167 Simulus spp. 799 NA      NA     NA    NA    NA     Simulus
#> 5  5479    5479 Simulus spp. 799 NA      NA     NA    NA    NA     Simulus
#> # … with 22 more variables: main_common_name <chr>, authority <chr>,
#> #   published_year <chr>, assessment_date <chr>, category <chr>,
#> #   criteria <chr>, population_trend <chr>, marine_system <chr>,
#> #   freshwater_system <chr>, terrestrial_system <chr>, assessor <chr>,
#> #   reviewer <chr>, aoo_km2 <chr>, eoo_km2 <chr>, elevation_upper <dbl>,
#> #   elevation_lower <dbl>, depth_upper <dbl>, depth_lower <dbl>,
#> #   errata_flag <chr>, errata_reason <chr>, amended_flag <chr>, …
# }