Simulate species data

Simulate species data for creating Area of Habitat data (Brooks et al. 2019). Specifically, data are simulated to define species geographic ranges, summary information, and habitat preferences.

simulate_spp_data(
  n,
  boundary_data,
  habitat_data = NULL,
  elevation_data = NULL,
  crosswalk_data = NULL,
  rf_model_scale_min = 1e+05,
  rf_model_scale_max = 2e+05,
  cache_dir = tempdir(),
  habitat_version = "latest",
  force = FALSE,
  omit_habitat_codes = iucn_habitat_codes_marine(),
  verbose = TRUE
)

Arguments

n
integer Number of species to simulate.

boundary_data
sf::st_sf() Spatial object delineating the spatial extent and boundary for simulating species ranges.

habitat_data
terra::rast() Raster data indicating the presence of different habitat classes across world (e.g. Jung et al. 2020a,b). Each pixel should contain an integer value that specifies which habitat class is present within the pixel (based on the argument to crosswalk_data). Defaults to NULL such that data are automatically obtained (using get_global_habitat_data()).

elevation_data
terra::rast() Raster data delineating the worldwide elevation data (e.g. Robinson et al. 2014). Defaults to NULL such that data are automatically obtained (using get_global_elevation_data()). If the data are obtained automatically, then a preprocessed version of the habitat data will be used to reduce processing time.

crosswalk_data
data.frame() Table containing data that indicate which pixel values in the argument to habitat_data correspond to which IUCN habitat classification codes. The argument should contain a code column that specifies a set of IUCN habitat classification codes (see iucn_habitat_data(), and a value column that specifies different values in the argument to habitat_data. Defaults to NULL such that the crosswalk for the default habitat data are used (i.e. crosswalk_jung_data()).

rf_model_scale_min
numeric Minimum scaling parameter used to control the smallest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 1e+5.

rf_model_scale_max
numeric Maximum scaling parameter used to control the largest possible size of simulated species ranges. See the documentation for the scale parameter in RandomFields::RMgauss() for more information. Defaults to 2e+5.

cache_dir
character folder path for downloading and caching data. By default, a temporary directory is used (i.e. tempdir()). To avoid downloading the same data multiple times, it is strongly recommended to specify a persistent storage location (see Examples below).

habitat_version
character Version of the habitat dataset that should be used. See documentation for the the version parameter in the get_global_habitat_data() function for further details. This parameter is only used if habitat data are obtained automatically (i.e. the argument to habitat_data is NULL). Defaults to "latest" such that the most recent version of the dataset is used if data need to be obtained.

force
logical should the data be downloaded even if the the data are already available? Defaults to FALSE.

omit_habitat_codes
character Habitat classification codes to omit from resulting Area of Habitat data. Please see the IUCN Red List Habitat Classification Scheme for the full range of habitat classification codes. For example, if the aim is to identify natural places that contain suitable conditions, then areas classified as anthropogenically modified (iucn_habitat_codes_artificial()), introduced vegetation (iucn_habitat_codes_introduced(), or unknown habitat (iucn_habitat_codes_misc()) should be excluded. Defaults to iucn_habitat_codes_marine(), such that marine habitats are excluded.

verbose
logical Should progress be displayed while processing data? Defaults to TRUE.

Value

A list object containing simulated data that are formatted following conventions used by the International Union for Conservation of Nature (IUCN) Red List of Threatened Species. It contains the following elements:

spp_range_data

A sf::st_sf() object containing the species' geographic range data.

spp_summary_data

A tibble::tibble() object containing summary information about the species (including altitudinal limit information.

spp_habitat_data

A tibble::tibble() object containing habitat preferences for the species.

References

Brooks TM, Pimm SL, Akçakaya HR, Buchanan GM, Butchart SHM, Foden W, Hilton-Taylor C, Hoffmann M, Jenkins CN, Joppa L, Li BV, Menon V, Ocampo-Peñuela N, Rondinini C (2019) Measuring terrestrial Area of Habitat (AOH) and its Utility for the IUCN Red List. Trends in Ecology & Evolution, 34:977--986. Available at https://doi.org/10.1016/j.tree.2019.06.009.

See also

Examples

# please ensure that the RandomFields and smoothr packages are installed
# to run these examples

# \dontrun{
# define persistent storage location
download_dir <- rappdirs::user_data_dir("aoh")

# create download directory if needed
if (!file.exists(download_dir)) {
  dir.create(download_dir, showWarnings = FALSE, recursive = TRUE)
}

# specify file path for boundary data
boundary_path <- system.file("shape/nc.shp", package = "sf")

# import boundary data to simulate species data
boundary_data <- sf::st_union(sf::read_sf(boundary_path))

# set random number generator seeds for consistency
set.seed(500)
RandomFields::RFoptions(seed = 500)

# simulate data for 5 species
x <- simulate_spp_data(
  n = 5, boundary_data = boundary_data, cache_dir = download_dir
)
#> ℹ importing global elevation data
#> ✔ importing global elevation data [11.8s]
#> 
#> ℹ importing global habitat data
#> Error: x does not inherit from class sf
#> ✖ importing global habitat data [11s]
#> 

# preview species range data
print(x$spp_range_data)
#> Error in print(x$spp_range_data) object 'x' not found

# preview species habitat affiliation data
print(x$spp_habitat_data)
#> Error in print(x$spp_habitat_data) object 'x' not found

# preview species summary data
print(x$spp_summary_data)
#> Error in print(x$spp_summary_data) object 'x' not found
# }