Project a raster using GDAL

This function is a wrapper for gdalUtils::gdalwarp().

terra_gdal_project(
  x,
  y,
  method = "bilinear",
  parallel_n_threads = 1,
  filename = tempfile(fileext = ".tif"),
  verbose = TRUE,
  ...
)

Arguments

x
terra::rast() Raster object with source data.

y
terra::rast() Raster object specifying spatial properties for output data.

method
character Name of interpolation method.

parallel_n_threads
integer Number of computational threads to use for data processing. To reduce run time, it is strongly recommended to set this parameter based on available resources (see Examples section below). Defaults to 1.

filename
character Filename for output raster. Defaults to tempfile(fileext = ".tif").

verbose
logical should information be displayed during processing? Defaults to TRUE.

...
Arguments passed to terra::writeRaster().

Value

A terra::rast() raster object.

Examples

# please ensure that the gdalUtils package is installed and
# GDAL system binaries are installed to run this example

# create raster with data
x <- rast(
  ncols = 40, nrows = 40, xmin = -110, xmax = -90, ymin = 40, ymax=60,
  crs = "+proj=longlat +datum=WGS84"
)
values(x) <- 1:ncell(x)

# create raster to define spatial properties for projection
y <- rast(
  ncols = 94, nrows = 124, xmin = -944881, xmax = 935118, ymin = 4664377,
  ymax = 7144377,
  crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
)

# project data
z <- terra_gdal_project(x, y)
#> Checking gdal_installation...
#> Scanning for GDAL installations...
#> Checking Sys.which...
#> GDAL version 3.2.2
#> GDAL command being used: "/usr/bin/gdalwarp" -overwrite  -te -944881 4664377 935118 7144377 -te_srs "/tmp/RtmpUAUXnt/file20425bb87b07.wkt" -tr 19999.9893617021 20000 -s_srs "/tmp/RtmpUAUXnt/file20421d6ed29e.wkt" -t_srs "/tmp/RtmpUAUXnt/file20425bb87b07.wkt" -r "bilinear" -of "GTiff" -wo "NUM_THREADS=1" -co "COMPRESS=LZW" "/tmp/RtmpUAUXnt/file204274abc8db.tif" "/tmp/RtmpUAUXnt/file20425b501a00.tif"

# preview result
print(z)
#> class       : SpatRaster 
#> dimensions  : 124, 94, 1  (nrow, ncol, nlyr)
#> resolution  : 19999.99, 20000  (x, y)
#> extent      : -944881, 935118, 4664377, 7144377  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=lcc +lat_0=0 +lon_0=-100 +lat_1=48 +lat_2=33 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> source      : file20425b501a00.tif 
#> name        : lyr.1